refactor: 구간추가 함수 위치 변경
구간 추가 함수 위치 변경
- 기존 Line에서 Sections의 책임으로 변경
