+++ src/test/java/nextstep/subway/line/LineTest.java
@@ -62,11 +62,9 @@ public class LineTest {
     @Test
     void addDuplicationSection() {
         Section 양재_수지구청역 = new Section(신분당선, 양재역, 수지구청역, 30);
-        boolean isUpStationExist = 신분당선.isStationExist(양재_수지구청역.getUpStation());
-        boolean isDownStationExist = 신분당선.isStationExist(양재_수지구청역.getDownStation());
 
         assertThatThrownBy(() -> {
-            신분당선.validSection(isUpStationExist, isDownStationExist);
+            신분당선.addSection(양재_수지구청역);
         }).isInstanceOf(RuntimeException.class)
                 .hasMessageContaining("동일한 구간은 추가할 수 없습니다.");
     }
@@ -74,12 +72,10 @@ public class LineTest {
     @DisplayName("상/하행선 둘중 하나와도 일치하지 않는 구간 추가 불가")
     @Test
     void addSameOneSection() {
-        Section 양재_수지구청역 = new Section(신분당선, 양재숲역, 판교역, 30);
-        boolean isUpStationExist = 신분당선.isStationExist(양재_수지구청역.getUpStation());
-        boolean isDownStationExist = 신분당선.isStationExist(양재_수지구청역.getDownStation());
+        Section 양재숲_판교역 = new Section(신분당선, 양재숲역, 판교역, 30);
 
         assertThatThrownBy(() -> {
-            신분당선.validSection(isUpStationExist, isDownStationExist);
+            신분당선.addSection(양재숲_판교역);
         }).isInstanceOf(RuntimeException.class)
                 .hasMessageContaining("상/하행선 둘 중 하나는 일치해야 합니다.");
     }
